+++ src/pages/Manager/ManagedApplicationList/index.tsx
@@ -1,124 +1,22 @@
-import type { MouseEvent } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
-import type { ClubApplicationsResponse } from '@/apis/club/entity';
-import CheckIcon from '@/assets/svg/check.svg';
-import CloseIcon from '@/assets/svg/close.svg';
-import PersonIcon from '@/assets/svg/person.svg';
-import { useToastContext } from '@/contexts/useToastContext';
-import {
-  useApproveManagedApplicationMutation,
-  useRejectManagedApplicationMutation,
-} from '@/pages/Manager/hooks/useManagedApplicationMutations';
+import { useParams } from 'react-router-dom';
 import { useGetManagedApplications } from '@/pages/Manager/hooks/useManagedApplications';
 import ManagerInfoCard from '@/pages/User/MyPage/components/ManagerInfoCard';
-import { useApiErrorToast } from '@/utils/hooks/error/useApiErrorToast';
 import { useInfiniteScroll } from '@/utils/hooks/useInfiniteScroll';
-import { formatIsoDateToYYYYMMDDHHMM } from '@/utils/ts/datetime/date';
-
-type ManagedApplication = ClubApplicationsResponse['applications'][number];
-
-function ApplicationAvatar({ imageUrl, name }: Pick<ManagedApplication, 'imageUrl' | 'name'>) {
-  if (imageUrl) {
-    return <img className="h-10 w-10 rounded-[10px] object-cover" src={imageUrl} alt={`${name} 프로필 이미지`} />;
-  }
-
-  return (
-    <div className="bg-indigo-25 flex h-10 w-10 items-center justify-center rounded-[10px] text-indigo-400">
-      <PersonIcon className="h-6 w-6" />
-    </div>
-  );
-}
-
-interface ApplicationCardProps {
-  application: ManagedApplication;
-  disabled: boolean;
-  onApprove: (e: MouseEvent<HTMLButtonElement>, applicationId: number) => void;
-  onReject: (e: MouseEvent<HTMLButtonElement>, applicationId: number) => void;
-  onDetail: (applicationId: number) => void;
-}
-
-function ApplicationCard({ application, disabled, onApprove, onReject, onDetail }: ApplicationCardProps) {
-  return (
-    <div
-      className="border-indigo-5 active:bg-indigo-5/60 flex cursor-pointer items-center justify-between rounded-2xl border bg-white p-3"
-      onClick={() => onDetail(application.id)}
-    >
-      <div className="flex min-w-0 flex-1 items-center gap-3">
-        <ApplicationAvatar imageUrl={application.imageUrl} name={application.name} />
-        <div className="min-w-0">
-          <div className="truncate text-[15px] leading-6 font-semibold text-indigo-700">
-            {application.name} ({application.studentNumber})
-          </div>
-          <div className="text-[13px] leading-[1.6] font-medium text-indigo-300">
-            지원일 : {formatIsoDateToYYYYMMDDHHMM(application.appliedAt)}
-          </div>
-        </div>
-      </div>
-
-      <div className="ml-3 flex shrink-0 items-center gap-3">
-        <button
-          type="button"
-          onClick={(e) => onApprove(e, application.id)}
-          disabled={disabled}
-          aria-label={`${application.name} 지원 승인`}
-          className="flex h-6 w-6 items-center justify-center text-[#69BFDF] disabled:opacity-50"
-        >
-          <CheckIcon className="size-6.5" />
-        </button>
-
-        <button
-          type="button"
-          onClick={(e) => onReject(e, application.id)}
-          disabled={disabled}
-          aria-label={`${application.name} 지원 거절`}
-          className="flex h-6 w-6 items-center justify-center text-indigo-300 disabled:opacity-50"
-        >
-          <CloseIcon className="size-6" />
-        </button>
-      </div>
-    </div>
-  );
-}
+import ApplicationCard from './components/ApplicationCard';
+import { useManagedApplicationActions } from './hooks/useManagedApplicationActions';
 
 function ManagedApplicationList() {
   const params = useParams();
-  const navigate = useNavigate();
   const clubId = Number(params.clubId);
-  const { showToast } = useToastContext();
-  const showApiErrorToast = useApiErrorToast();
-
-  const limit = 10;
 
   const { managedClubApplicationList, applications, fetchNextPage, hasNextPage, isFetchingNextPage, hasNoRecruitment } =
-    useGetManagedApplications(clubId, { limit });
+    useGetManagedApplications(clubId, { limit: 10 });
 
-  const { mutate: approve, isPending: isApproving } = useApproveManagedApplicationMutation(clubId);
-  const { mutate: reject, isPending: isRejecting } = useRejectManagedApplicationMutation(clubId);
+  const { isPending, handleApprove, handleReject, handleDetail, handleChat } = useManagedApplicationActions(clubId);
+
   const observerRef = useInfiniteScroll(fetchNextPage, hasNextPage, isFetchingNextPage, {
     enabled: !hasNoRecruitment,
   });
-
-  const isPending = isApproving || isRejecting;
-
-  const handleApprove = (e: MouseEvent<HTMLButtonElement>, applicationId: number) => {
-    e.stopPropagation();
-    approve(applicationId, {
-      onSuccess: () => showToast('지원이 승인되었습니다'),
-      onError: (error) => showApiErrorToast(error, '지원 승인 처리에 실패했습니다.'),
-    });
-  };
-
-  const handleReject = (e: MouseEvent<HTMLButtonElement>, applicationId: number) => {
-    e.stopPropagation();
-    reject(applicationId, {
-      onSuccess: () => showToast('지원이 거절되었습니다'),
-      onError: (error) => showApiErrorToast(error, '지원 거절 처리에 실패했습니다.'),
-    });
-  };
-
-  const handleDetail = (applicationId: number) => {
-    navigate(`${applicationId}`);
-  };
 
   if (hasNoRecruitment) {
     return (
@@ -151,6 +49,7 @@ function ManagedApplicationList() {
               onApprove={handleApprove}
               onReject={handleReject}
               onDetail={handleDetail}
+              onChat={handleChat}
             />
           ))
         ) : (
